Move sedes inference functions
Necessary to avoid circular import
